Implement Migrate shell feature to ADR-0001 structure (#194)
