+++ apps/carbotracker/src/features/shell/+state/actions/component.actions.ts
@@ -1,0 +1,11 @@
+import { createActionGroup, emptyProps } from '@ngrx/store';
+
+export const ShellComponentActions = createActionGroup({
+  source: 'Shell | Shell Component',
+  events: {
+    'Products Clicked': emptyProps(),
+    'Current Meal Clicked': emptyProps(),
+    'Saved Meals Clicked': emptyProps(),
+    'Settings Clicked': emptyProps(),
+  },
+});
