+++ apps/carbotracker/src/features/shell/+state/index.ts
@@ -1,0 +1,4 @@
+export * from './actions/component.actions';
+export * from './actions/routing.actions';
+
+export * as routingEffects from './effects/routing.effects';
