--- apps/carbotracker/src/shell-feature/shell.effects.ts
@@ -1,90 +1,0 @@
-import { inject } from '@angular/core';
-import { Router } from '@angular/router';
-import { Actions, createEffect, ofType } from '@ngrx/effects';
-import { catchError, from, map, of, switchMap } from 'rxjs';
-import { PasswordApiActions } from '../features/auth/+state';
-import {
-  AccountPageActions,
-  InsulinToCarbRatiosPageActions,
-} from '../features/settings/+state';
-import {
-  ShellComponentActions,
-  ShellRouterEffectsActions,
-} from './shell.actions';
-
-export const navigateToProducts$ = createEffect(
-  (actions$ = inject(Actions), router = inject(Router)) =>
-    actions$.pipe(
-      ofType(ShellComponentActions.productsClicked),
-      switchMap(() =>
-        from(router.navigate(['app', 'products'])).pipe(
-          map(() =>
-            ShellRouterEffectsActions.navigationToProductsPageSuccessful(),
-          ),
-          catchError(() =>
-            of(ShellRouterEffectsActions.navigationToProductsPageFailed()),
-          ),
-        ),
-      ),
-    ),
-  { functional: true },
-);
-
-export const navigateToCurrentMeal$ = createEffect(
-  (actions$ = inject(Actions), router = inject(Router)) =>
-    actions$.pipe(
-      ofType(ShellComponentActions.currentMealClicked),
-      switchMap(() =>
-        from(router.navigate(['app', 'current-meal'])).pipe(
-          map(() =>
-            ShellRouterEffectsActions.navigationToCurrentMealPageSuccessful(),
-          ),
-          catchError(() =>
-            of(ShellRouterEffectsActions.navigationToCurrentMealPageFailed()),
-          ),
-        ),
-      ),
-    ),
-  { functional: true },
-);
-
-export const navigateToSavedMeals$ = createEffect(
-  (actions$ = inject(Actions), router = inject(Router)) =>
-    actions$.pipe(
-      ofType(ShellComponentActions.savedMealsClicked),
-      switchMap(() =>
-        from(router.navigate(['app', 'saved-meals'])).pipe(
-          map(() =>
-            ShellRouterEffectsActions.navigationToSavedMealsPageSuccessful(),
-          ),
-          catchError(() =>
-            of(ShellRouterEffectsActions.navigationToSavedMealsPageFailed()),
-          ),
-        ),
-      ),
-    ),
-  { functional: true },
-);
-
-export const navigateToSettingsPage$ = createEffect(
-  (actions$ = inject(Actions), router = inject(Router)) =>
-    actions$.pipe(
-      ofType(
-        ShellComponentActions.settingsClicked,
-        AccountPageActions.goBackIconClicked,
-        PasswordApiActions.updatePasswordSuccessful,
-        InsulinToCarbRatiosPageActions.goBackIconClicked,
-      ),
-      switchMap(() =>
-        from(router.navigate(['app', 'settings'])).pipe(
-          map(() =>
-            ShellRouterEffectsActions.navigationToSettingsPageSuccessful(),
-          ),
-          catchError(() =>
-            of(ShellRouterEffectsActions.navigationToSettingsPageFailed()),
-          ),
-        ),
-      ),
-    ),
-  { functional: true },
-);
